refactor: 인증 관련 클래스를 auth 모듈로 이동
