+++ server/src/main/java/com/emr/slgi/auth/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.emr.slgi.config;
+package com.emr.slgi.auth.config;
 
 import java.util.Map;
