+++ server/src/main/java/com/emr/slgi/auth/service/RefreshTokenService.java
@@ -13,7 +13,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
-import com.emr.slgi.util.JwtUtil;
+import com.emr.slgi.auth.util.JwtUtil;
 
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
